# docs/api.yml
openapi: "3.1.0"
info:
  title: Grimoire
  version: "1.0.0"
  description: |
    Self-hosted TTRPG library manager.

    All endpoints except `/api/auth/status`, `/api/auth/setup`, and `/api/auth/login`
    require a valid JWT passed as `Authorization: Bearer <token>`, or as a `?token=`
    query parameter (required for browser-embedded images and file downloads).

    **Roles:**
    - `admin` — full access including user management and app settings
    - `gm` — read everything, edit metadata, rescan library, manage campaigns
    - `player` — read-only access
servers:
  - url: http://localhost:9481
    description: Local instance
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tokenParam:
      type: apiKey
      in: query
      name: token
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - bearerAuth: []
  - tokenParam: []
paths:
...
  /api/books/{id}/toc:
    get:
      tags: [books]
      summary: PDF table of contents
      description: Returns the PDF outline as nested `{title, page, level, children}` entries.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: TOC
          content:
            application/json:
              schema:
                type: object
                properties:
                  toc:
                    type: array
                    items: { type: object }
  /api/books/{id}/page/{num}:
    get:
      tags: [books]
      summary: Render a PDF page as WebP
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: num, required: true, schema: { type: integer } }
        - { in: query, name: width, schema: { type: integer, default: 1200, maximum: 3000 } }
      responses:
        "200": { description: WebP image }
        "400": { description: Page out of range }
        "404": { description: Not found }
  /api/books/{id}/page/{num}/text:
    get:
      tags: [books]
      summary: Get page text
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: num, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Page text
          content:
            application/json:
              schema:
                type: object
                properties:
                  text: { type: string }
  /api/books/{id}/page/{num}/words:
    get:
      tags: [books]
      summary: Get page word bounding boxes
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: num, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Word bounding boxes
          content:
            application/json:
              schema:
                type: object
                properties:
                  width: { type: number }
                  height: { type: number }
                  words:
                    type: array
                    items:
                      type: object
                      properties:
                        x0: { type: number }
                        y0: { type: number }
                        x1: { type: number }
                        y1: { type: number }
                        text: { type: string }

  # ── Maps ──────────────────────────────────────────────────────────────────
...